+++ common/src/main/java/com/faboslav/structurify/common/mixin/LocateCommandAccessor.java
@@ -1,0 +1,5 @@
+package com.faboslav.structurify.common.mixin;
+
+public class LocateCommandAccessor
+{
+}